test: preserve fast contracts in unit tier
