+++ pkg/cluster/node_channel_facade_integration_test.go
@@ -4,7 +4,6 @@ package cluster
 
 import (
 	"context"
-	"errors"
 	"testing"
 
 	channelruntime "github.com/WuKongIM/WuKongIM/pkg/channel"
@@ -91,23 +90,5 @@ func TestNodeChannelRuntimeEvictDelegatesToChannels(t *testing.T) {
 	}
 	if runtime.evictCalls != 1 || len(runtime.lastEvict.ChannelIDs) != 1 || runtime.lastEvict.ChannelIDs[0] != channelID {
 		t.Fatalf("evict calls=%d selector=%#v, want selector forwarded", runtime.evictCalls, runtime.lastEvict)
-	}
-}
-
-func TestNodeChannelRuntimeRequiresStartedChannels(t *testing.T) {
-	node, err := New(validNodeConfig(t))
-	if err != nil {
-		t.Fatalf("New() error = %v", err)
-	}
-	node.started.Store(true)
-
-	if _, err := node.ChannelRuntimeSnapshot(context.Background()); !errors.Is(err, ErrNotStarted) {
-		t.Fatalf("ChannelRuntimeSnapshot() error = %v, want ErrNotStarted", err)
-	}
-	if _, err := node.ChannelRuntimeProbe(context.Background(), channelruntime.RuntimeSelector{}); !errors.Is(err, ErrNotStarted) {
-		t.Fatalf("ChannelRuntimeProbe() error = %v, want ErrNotStarted", err)
-	}
-	if _, err := node.ChannelRuntimeEvict(context.Background(), channelruntime.RuntimeSelector{}); !errors.Is(err, ErrNotStarted) {
-		t.Fatalf("ChannelRuntimeEvict() error = %v, want ErrNotStarted", err)
 	}
 }
