+++ pkg/cluster/node_meta_channel_scan_integration_test.go
