+++ pkg/cluster/node_meta_user_scan_integration_test.go
