+++ pkg/slot/fsm/channel_migration_cmds_integration_test.go
@@ -12,173 +12,6 @@ import (
 	"github.com/WuKongIM/WuKongIM/pkg/slot/multiraft"
 )
 
-func TestChannelMigrationCommandCodecRoundTrip(t *testing.T) {
-	task := fsmTestChannelMigrationTask("task-codec", "channel-codec")
-	task.Status = metadb.ChannelMigrationStatusRunning
-	task.Phase = metadb.ChannelMigrationPhaseWarmCatchUp
-	task.OwnerNodeID = 1
-	task.OwnerLeaseUntilMS = 1750000005000
-	task.UpdatedAtMS = 1750000001000
-	advance := fsmTestChannelMigrationAdvance(task, metadb.ChannelMigrationStatusRunning, metadb.ChannelMigrationPhaseFinalTargetCatchUp, 1750000002000)
-	advance.CutoverProof = metadb.ChannelMigrationCutoverProof{
-		CutoverLEO:               100,
-		CutoverHW:                99,
-		DrainedLeaderNode:        1,
-		DrainedRuntimeGeneration: 2,
-		DrainedChannelEpoch:      3,
-		DrainedLeaderEpoch:       4,
-		DrainedFenceVersion:      7,
-	}
-
-	cases := []struct {
-		name string
-		data []byte
-		want command
-	}{
-		{
-			name: "create",
-			data: EncodeCreateChannelMigrationTaskCommand(fsmTestChannelMigrationTask("task-create", "channel-create")),
-			want: &createChannelMigrationTaskCmd{task: fsmTestChannelMigrationTask("task-create", "channel-create")},
-		},
-		{
-			name: "create_with_runtime_guard",
-			data: EncodeCreateChannelMigrationTaskWithRuntimeGuardCommand(metadb.ChannelMigrationTaskCreate{
-				Task:         fsmTestChannelMigrationTask("task-create-guard", "channel-create-guard"),
-				RuntimeGuard: fsmTestRuntimeGuard(fsmTestRuntimeMeta("channel-create-guard", 1)),
-			}),
-			want: &createChannelMigrationTaskWithRuntimeGuardCmd{req: metadb.ChannelMigrationTaskCreate{
-				Task:         fsmTestChannelMigrationTask("task-create-guard", "channel-create-guard"),
-				RuntimeGuard: fsmTestRuntimeGuard(fsmTestRuntimeMeta("channel-create-guard", 1)),
-			}},
-		},
-		{
-			name: "claim",
-			data: EncodeClaimChannelMigrationTaskCommand(fsmTestChannelMigrationClaim(task, 2, 1750000006000, 1750000002000)),
-			want: &claimChannelMigrationTaskCmd{req: fsmTestChannelMigrationClaim(task, 2, 1750000006000, 1750000002000)},
-		},
-		{
-			name: "advance",
-			data: EncodeAdvanceChannelMigrationTaskCommand(advance),
-			want: &advanceChannelMigrationTaskCmd{req: advance},
-		},
-		{
-			name: "set_fence",
-			data: EncodeSetChannelWriteFenceCommand(fsmTestSetFenceRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000007000, 1750000002000)),
-			want: &setChannelWriteFenceCmd{req: fsmTestSetFenceRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000007000, 1750000002000)},
-		},
-		{
-			name: "reset_fence",
-			data: EncodeResetChannelWriteFenceToPreCutoverCommand(fsmTestResetFenceRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), metadb.ChannelMigrationPhaseWarmCatchUp, 1750000003000)),
-			want: &resetChannelWriteFenceToPreCutoverCmd{req: fsmTestResetFenceRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), metadb.ChannelMigrationPhaseWarmCatchUp, 1750000003000)},
-		},
-		{
-			name: "commit_leader",
-			data: EncodeCommitChannelLeaderTransferCommand(fsmTestCommitLeaderRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)),
-			want: &commitChannelLeaderTransferCmd{req: fsmTestCommitLeaderRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)},
-		},
-		{
-			name: "add_learner",
-			data: EncodeAddChannelLearnerCommand(fsmTestAddLearnerRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000003000)),
-			want: &addChannelLearnerCmd{req: fsmTestAddLearnerRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000003000)},
-		},
-		{
-			name: "promote",
-			data: EncodePromoteLearnerAndRemoveReplicaCommand(fsmTestPromoteRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)),
-			want: &promoteLearnerAndRemoveReplicaCmd{req: fsmTestPromoteRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)},
-		},
-		{
-			name: "clear_fence",
-			data: EncodeClearChannelWriteFenceCommand(fsmTestClearFenceRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)),
-			want: &clearChannelWriteFenceCmd{req: fsmTestClearFenceRequest(task, fsmTestFencedRuntimeMeta(task.ChannelID, task.ChannelType, task.TaskID, 7), 1750000003000)},
-		},
-		{
-			name: "abort",
-			data: EncodeAbortChannelMigrationCommand(fsmTestAbortRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000003000)),
-			want: &abortChannelMigrationCmd{req: fsmTestAbortRequest(task, fsmTestRuntimeMeta(task.ChannelID, task.ChannelType), 1750000003000)},
-		},
-		{
-			name: "garbage_collect",
-			data: EncodeGarbageCollectTerminalChannelMigrationTasksCommand(metadb.ChannelMigrationTaskGCRequest{BeforeMS: 1750000010000, Limit: 10}),
-			want: &garbageCollectMigrationTasksCmd{req: metadb.ChannelMigrationTaskGCRequest{BeforeMS: 1750000010000, Limit: 10}},
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.name, func(t *testing.T) {
-			got, err := decodeCommand(tc.data)
-			if err != nil {
-				t.Fatalf("decodeCommand(%s) error = %v", tc.name, err)
-			}
-			if !reflect.DeepEqual(got, tc.want) {
-				t.Fatalf("decodeCommand(%s) = %#v, want %#v", tc.name, got, tc.want)
-			}
-		})
-	}
-}
-
-func TestChannelMigrationCommandRejectsDuplicatePayload(t *testing.T) {
-	data := []byte{commandVersion, cmdTypeCreateChannelMigrationTask}
-	data = appendBytesTLVField(data, tagChannelMigrationCommandPayload, []byte(`{}`))
-	data = appendBytesTLVField(data, tagChannelMigrationCommandPayload, []byte(`{}`))
-
-	_, err := decodeCommand(data)
-	if err == nil {
-		t.Fatal("decodeCommand(duplicate payload) error = nil, want error")
-	}
-	if !errors.Is(err, metadb.ErrCorruptValue) {
-		t.Fatalf("decodeCommand(duplicate payload) error = %v, want ErrCorruptValue", err)
-	}
-}
-
-func TestChannelMigrationCommandRejectsAmbiguousJSONPayload(t *testing.T) {
-	tests := []struct {
-		name    string
-		payload string
-	}{
-		{
-			name:    "unknown_field",
-			payload: `{"TaskID":"task-json","UnknownField":true}`,
-		},
-		{
-			name:    "duplicate_key",
-			payload: `{"TaskID":"task-json","TaskID":"task-json-2"}`,
-		},
-		{
-			name:    "trailing_value",
-			payload: `{"TaskID":"task-json"} {}`,
-		},
-		{
-			name:    "case_insensitive_duplicate_key",
-			payload: `{"TaskID":"task-json","taskid":"task-json-2"}`,
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			data := []byte{commandVersion, cmdTypeCreateChannelMigrationTask}
-			data = appendBytesTLVField(data, tagChannelMigrationCommandPayload, []byte(tc.payload))
-
-			_, err := decodeCommand(data)
-			if err == nil {
-				t.Fatalf("decodeCommand(%s) error = nil, want error", tc.name)
-			}
-			if !errors.Is(err, metadb.ErrCorruptValue) {
-				t.Fatalf("decodeCommand(%s) error = %v, want ErrCorruptValue", tc.name, err)
-			}
-		})
-	}
-}
-
-func TestStateMachineStaleCommitErrorsAreDeterministicResults(t *testing.T) {
-	for _, err := range []error{metadb.ErrStaleMeta, metadb.ErrNotFound, metadb.ErrAlreadyExists} {
-		if !isStaleMetaCommitError(err) {
-			t.Fatalf("isStaleMetaCommitError(%v) = false, want true", err)
-		}
-	}
-	if isStaleMetaCommitError(metadb.ErrInvalidArgument) {
-		t.Fatal("isStaleMetaCommitError(ErrInvalidArgument) = true, want false")
-	}
-}
-
 func TestStateMachineChannelApplyBatchRuntimeMetaCreateThenAddLearner(t *testing.T) {
 	ctx := context.Background()
 	db := openTestDB(t)
@@ -1626,210 +1459,5 @@ func TestStateMachineChannelMigrationGarbageCollectDeletesTerminalTasks(t *testi
 	}
 	if gotNew.TaskID != newTask.TaskID {
 		t.Fatalf("new task id = %q, want %q", gotNew.TaskID, newTask.TaskID)
-	}
-}
-
-func fsmTestChannelMigrationTask(taskID, channelID string) metadb.ChannelMigrationTask {
-	return metadb.ChannelMigrationTask{
-		TaskID:           taskID,
-		Kind:             metadb.ChannelMigrationKindReplicaReplace,
-		Status:           metadb.ChannelMigrationStatusPending,
-		Phase:            metadb.ChannelMigrationPhaseValidate,
-		ChannelID:        channelID,
-		ChannelType:      1,
-		SourceNode:       2,
-		TargetNode:       3,
-		DesiredLeader:    1,
-		BaseChannelEpoch: 10,
-		BaseLeaderEpoch:  20,
-		CreatedAtMS:      1750000000000,
-		UpdatedAtMS:      1750000000000,
-	}
-}
-
-func fsmTestLeaderTransferTask(taskID, channelID string) metadb.ChannelMigrationTask {
-	task := fsmTestChannelMigrationTask(taskID, channelID)
-	task.Kind = metadb.ChannelMigrationKindLeaderTransfer
-	task.Status = metadb.ChannelMigrationStatusRunning
-	task.Phase = metadb.ChannelMigrationPhaseProbeTarget
-	task.SourceNode = 1
-	task.TargetNode = 2
-	task.DesiredLeader = 2
-	task.UpdatedAtMS = 1750000001000
-	return task
-}
-
-func fsmTestRuntimeMeta(channelID string, channelType int64) metadb.ChannelRuntimeMeta {
-	return metadb.ChannelRuntimeMeta{
-		ChannelID:    channelID,
-		ChannelType:  channelType,
-		ChannelEpoch: 10,
-		LeaderEpoch:  20,
-		Replicas:     []uint64{1, 2},
-		ISR:          []uint64{1, 2},
-		Leader:       1,
-		MinISR:       2,
-		Status:       1,
-		Features:     1,
-		LeaseUntilMS: 1750000010000,
-	}
-}
-
-func fsmTestFencedRuntimeMeta(channelID string, channelType int64, taskID string, version uint64) metadb.ChannelRuntimeMeta {
-	meta := fsmTestRuntimeMeta(channelID, channelType)
-	meta.WriteFenceToken = taskID
-	meta.WriteFenceVersion = version
-	meta.WriteFenceReason = 1
-	meta.WriteFenceUntilMS = 1750000010000
-	return meta
-}
-
-func fsmTestTaskGuard(task metadb.ChannelMigrationTask) metadb.ChannelMigrationTaskGuard {
-	return metadb.ChannelMigrationTaskGuard{
-		ChannelID:                 task.ChannelID,
-		ChannelType:               task.ChannelType,
-		TaskID:                    task.TaskID,
-		ExpectedStatus:            task.Status,
-		ExpectedPhase:             task.Phase,
-		ExpectedOwnerNodeID:       task.OwnerNodeID,
-		ExpectedOwnerLeaseUntilMS: task.OwnerLeaseUntilMS,
-		ExpectedUpdatedAtMS:       task.UpdatedAtMS,
-	}
-}
-
-func fsmTestRuntimeGuard(meta metadb.ChannelRuntimeMeta) metadb.ChannelMigrationRuntimeGuard {
-	return metadb.ChannelMigrationRuntimeGuard{
-		ChannelID:            meta.ChannelID,
-		ChannelType:          meta.ChannelType,
-		ExpectedChannelEpoch: meta.ChannelEpoch,
-		ExpectedLeaderEpoch:  meta.LeaderEpoch,
-		ExpectedLeader:       meta.Leader,
-		ExpectedFenceToken:   meta.WriteFenceToken,
-		ExpectedFenceVersion: meta.WriteFenceVersion,
-	}
-}
-
-func fsmTestChannelMigrationClaim(task metadb.ChannelMigrationTask, owner uint64, leaseUntilMS, updatedAtMS int64) metadb.ChannelMigrationTaskClaim {
-	return metadb.ChannelMigrationTaskClaim{
-		Guard:             fsmTestTaskGuard(task),
-		Status:            metadb.ChannelMigrationStatusRunning,
-		Phase:             task.Phase,
-		OwnerNodeID:       owner,
-		OwnerLeaseUntilMS: leaseUntilMS,
-		NowMS:             updatedAtMS,
-		UpdatedAtMS:       updatedAtMS,
-	}
-}
-
-func fsmTestChannelMigrationAdvance(task metadb.ChannelMigrationTask, status metadb.ChannelMigrationStatus, phase metadb.ChannelMigrationPhase, updatedAtMS int64) metadb.ChannelMigrationTaskAdvance {
-	return metadb.ChannelMigrationTaskAdvance{
-		Guard:       fsmTestTaskGuard(task),
-		Status:      status,
-		Phase:       phase,
-		UpdatedAtMS: updatedAtMS,
-	}
-}
-
-func fsmTestSetFenceRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, fenceUntilMS, updatedAtMS int64) metadb.ChannelMigrationFenceRequest {
-	return metadb.ChannelMigrationFenceRequest{
-		Guard:        fsmTestTaskGuard(task),
-		RuntimeGuard: fsmTestRuntimeGuard(meta),
-		Status:       metadb.ChannelMigrationStatusRunning,
-		Phase:        metadb.ChannelMigrationPhaseCutoverFence,
-		FenceReason:  1,
-		FenceUntilMS: fenceUntilMS,
-		UpdatedAtMS:  updatedAtMS,
-	}
-}
-
-func fsmTestResetFenceRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, phase metadb.ChannelMigrationPhase, updatedAtMS int64) metadb.ChannelMigrationResetFenceRequest {
-	return metadb.ChannelMigrationResetFenceRequest{
-		Guard:        fsmTestTaskGuard(task),
-		RuntimeGuard: fsmTestRuntimeGuard(meta),
-		Status:       metadb.ChannelMigrationStatusRunning,
-		Phase:        phase,
-		NowMS:        meta.WriteFenceUntilMS + 1,
-		UpdatedAtMS:  updatedAtMS,
-	}
-}
-
-func fsmTestCommitLeaderRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, updatedAtMS int64) metadb.ChannelMigrationLeaderTransferRequest {
-	return metadb.ChannelMigrationLeaderTransferRequest{
-		Guard:           fsmTestTaskGuard(task),
-		RuntimeGuard:    fsmTestRuntimeGuard(meta),
-		Status:          metadb.ChannelMigrationStatusRunning,
-		Phase:           metadb.ChannelMigrationPhaseVerifyNewLeader,
-		DesiredLeader:   task.TargetNode,
-		NextLeaderEpoch: meta.LeaderEpoch + 1,
-		LeaseUntilMS:    meta.LeaseUntilMS + 1000,
-		NowMS:           meta.WriteFenceUntilMS - 1,
-		UpdatedAtMS:     updatedAtMS,
-	}
-}
-
-func fsmTestAddLearnerRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, updatedAtMS int64) metadb.ChannelMigrationAddLearnerRequest {
-	return metadb.ChannelMigrationAddLearnerRequest{
-		Guard:        fsmTestTaskGuard(task),
-		RuntimeGuard: fsmTestRuntimeGuard(meta),
-		Status:       metadb.ChannelMigrationStatusRunning,
-		Phase:        metadb.ChannelMigrationPhaseBootstrapTarget,
-		TargetNode:   task.TargetNode,
-		UpdatedAtMS:  updatedAtMS,
-	}
-}
-
-func fsmTestPromoteRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, updatedAtMS int64) metadb.ChannelMigrationPromoteLearnerRequest {
-	return metadb.ChannelMigrationPromoteLearnerRequest{
-		Guard:        fsmTestTaskGuard(task),
-		RuntimeGuard: fsmTestRuntimeGuard(meta),
-		Status:       metadb.ChannelMigrationStatusRunning,
-		Phase:        metadb.ChannelMigrationPhaseVerifyMembership,
-		SourceNode:   task.SourceNode,
-		TargetNode:   task.TargetNode,
-		NowMS:        meta.WriteFenceUntilMS - 1,
-		UpdatedAtMS:  updatedAtMS,
-	}
-}
-
-func fsmTestClearFenceRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, updatedAtMS int64) metadb.ChannelMigrationClearFenceRequest {
-	return metadb.ChannelMigrationClearFenceRequest{
-		Guard:         fsmTestTaskGuard(task),
-		RuntimeGuard:  fsmTestRuntimeGuard(meta),
-		Status:        metadb.ChannelMigrationStatusCompleted,
-		Phase:         metadb.ChannelMigrationPhaseClearFence,
-		UpdatedAtMS:   updatedAtMS,
-		CompletedAtMS: updatedAtMS,
-	}
-}
-
-func fsmTestAbortRequest(task metadb.ChannelMigrationTask, meta metadb.ChannelRuntimeMeta, updatedAtMS int64) metadb.ChannelMigrationAbortRequest {
-	return metadb.ChannelMigrationAbortRequest{
-		Guard:         fsmTestTaskGuard(task),
-		RuntimeGuard:  fsmTestRuntimeGuard(meta),
-		Status:        metadb.ChannelMigrationStatusAborted,
-		Phase:         task.Phase,
-		UpdatedAtMS:   updatedAtMS,
-		CompletedAtMS: updatedAtMS,
-		LastError:     "operator aborted",
-	}
-}
-
-func setFSMTestDrainProof(task *metadb.ChannelMigrationTask, fenceVersion uint64) {
-	task.CutoverLEO = 100
-	task.CutoverHW = 99
-	task.DrainedLeaderNode = 1
-	task.DrainedRuntimeGeneration = 2
-	task.DrainedChannelEpoch = task.BaseChannelEpoch
-	task.DrainedLeaderEpoch = task.BaseLeaderEpoch
-	task.DrainedFenceVersion = fenceVersion
-}
-
-func requireFSMStaleResult(t *testing.T, result []byte, err error) {
-	t.Helper()
-	if err != nil && !errors.Is(err, metadb.ErrStaleMeta) {
-		t.Fatalf("unexpected stale result error = %v", err)
-	}
-	if got := string(result); got != ApplyResultStaleMeta {
-		t.Fatalf("stale result = %q, want %q", got, ApplyResultStaleMeta)
 	}
 }
